ci: don't cache builds of the actual project
The reasoning is that each run should build the repository contents from
scratch, using a cache here risks not doing that.

Additionally, each change to the contents of this repo triggers a new
build with a new hash, so without ignoring these files the cache will
continue to grow unbounded each run.