broken: create the canvas on wasm
Lots of changes here while trying to debug, but I'm going to commit and
come back to this later.

Notably the window on linux will render and then close out after a few
loops of the event_loop. It's getting a DestroyNotify event from winit
linux/x11/event_processor.rs on line 551.